+++ packages/core/src/index.ts
@@ -52,7 +52,7 @@ export type {
 } from './redux/interfaces'
 export * as actions from './redux/actions'
 export * as Rooms from './rooms'
-export type { RoomSessionRecording } from './rooms'
+export type { RoomSessionRecording, RoomSessionPlayback } from './rooms'
 export const selectors = {
   ...sessionSelectors,
 }