+++ packages/core/src/rooms/index.ts
@@ -10,3 +10,4 @@ export interface BaseRoomInterface<
 
 export * from './methods'
 export * from './RoomSessionRecording'
+export * from './RoomSessionPlayback'