+++ packages/realtime-api/src/video/Video.ts
@@ -8,6 +8,7 @@ import {
   ConsumerContract,
   RoomSessionRecording,
   VideoRoomEventParams,
+  RoomSessionPlayback,
 } from '@signalwire/core'
 
 import { RealTimeVideoApiEvents } from '../types/video'
@@ -64,6 +65,7 @@ export type {
   RoomSessionFullState,
   RoomSessionMember,
   RoomSessionRecording,
+  RoomSessionPlayback,
 }
 
 /** @internal */